Fix: BuildConfig file not generated/disappearing
Apparently using `variant.sources.java.all` in AGP 7.2 causes BuildConfig files to disappear or not generate at all, this is very likely a bug in AGP 7.2 because it is working in AGP 7.3.0-beta01. For now this is a workaround that reintroduces code that utilises old non-public AGP code, but at least this code is working.

Fixes #54